node-client/requirements.txt
# See https://grpc.io/docs/tutorials/basic/python.html
grpcio-tools

tornado
